+++ .eslintrc.cjs
@@ -1,9 +1,15 @@
 module.exports = {
+    "extends": "eslint:recommended",
     "env": {
         "browser": true,
         "es2020": true
     },
     "extends": "eslint:recommended",
+    "globals": {
+        "browser": true,
+        "chrome": true,
+        "opr": true
+    },
     "parserOptions": {
         "sourceType": "module"
     },