+++ src/constants.js
@@ -28,7 +28,7 @@ export {
   getResultValue
 }
 
-const VERSION = '2.0.5';
+const VERSION = '2.0.6';
 
 /**
  * @constant RULESET